requirements.txt
opencv-python==4.11.0.86numpy==2.1.1Pillow==11.1.0ultralytics==8.3.102easyocr==1.7.2deep_sort_realtime==1.3.2